[cleanup] Organize codebase
